ref #630 - using modal now instead of popup, can't dismiss modal, though
