+++ src/app/app.module.ts
@@ -32,7 +32,6 @@ registerLocaleData(localeEn);
 import { Geolocation } from '@ionic-native/geolocation/ngx';
 import { Calendar } from '@ionic-native/calendar/ngx';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
-import { CampusMapFeatureModalComponent } from './campus-map-feature-modal/campus-map-feature-modal.component';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -58,8 +57,7 @@ export class IonicGestureConfig extends HammerGestureConfig {
 
 @NgModule({
   declarations: [
-    AppComponent,
-    CampusMapFeatureModalComponent
+    AppComponent
   ],
   entryComponents: [],
   imports: [
