--- src/app/campus-map-feature-modal/campus-map-feature-modal.component.html
@@ -1,3 +1,0 @@
-<p>
-  campus-map-feature-modal works!
-</p>
