--- src/app/campus-map-feature-modal/campus-map-feature-modal.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-campus-map-feature-modal',
-  templateUrl: './campus-map-feature-modal.component.html',
-  styleUrls: ['./campus-map-feature-modal.component.scss']
-})
-export class CampusMapFeatureModalComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
