+++ src/app/components/campus-map-feature-modal/campus-map-feature-modal.component.html
@@ -1,0 +1,11 @@
+<ion-header>
+  <ion-toolbar>
+    <ion-button slot="start" (click)="closeModal()" fill="clear">
+      <ion-icon style="color: black; font-size: x-large;" name="ios-arrow-back"></ion-icon>
+    </ion-button>
+    <ion-title>{{ feature.properties.Name }}</ion-title>
+  </ion-toolbar>
+</ion-header>
+<ion-content padding>
+  {{ feature.properties.description }}
+</ion-content>
